+++ meta_package_manager/managers/npm.py
@@ -28,7 +28,7 @@ if TYPE_CHECKING:
     from collections.abc import Iterator
     from typing import Any
 
-    from click_extra.testing import TArg, TNestedArgs
+    from click_extra.execution import TArg, TNestedArgs
 
     from ..package import Package