+++ tests/test_bar_plugin.py
@@ -25,7 +25,7 @@ from typing import ClassVar, cast
 import pytest
 from boltons.iterutils import flatten
 from click_extra.envvar import env_copy
-from click_extra.testing import args_cleanup
+from click_extra.execution import args_cleanup
 from extra_platforms.pytest import unless_macos
 
 from meta_package_manager import bar_plugin